Moved NetworkConnectionPool into a separate class file.
