+++ src/main/java/org/reaktivity/nukleus/kafka/internal/stream/ClientStreamFactoryBuilder.java
@@ -29,7 +29,7 @@ import org.reaktivity.nukleus.stream.StreamFactoryBuilder;
 public final class ClientStreamFactoryBuilder implements StreamFactoryBuilder
 {
     private final KafkaConfiguration config;
-    private final Long2ObjectHashMap<ClientStreamFactory.NetworkConnectionPool.NetworkConnection> correlations;
+    private final Long2ObjectHashMap<NetworkConnectionPool.NetworkConnection> correlations;
 
     private RouteManager router;
     private MutableDirectBuffer writeBuffer;
